ai-services/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.1.5
httpx==0.26.0
python-dotenv==1.0.0
redis==5.0.1
